backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
supabase>=2.3,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0
requests>=2.31,<3.0

gunicorn>=21.2,<22.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=4.1,<5.0
pytest-factoryboy>=2.6,<3.0
